requirements.txt
streamlit
langchain
openai
faiss-cpu
pypdf
python-dotenv
langchain-openai
langchain-community
langchain-text-splitters


